Added bear embers VFX
Also added a bear glow material that can be used if it doesn't break the main model, can be used for damage flashing. Issue is that it replaces the eye, nose and finger materials even though those are still on the bear. Also the embers don't glow like they're supposed to because something is up with how URP is making the bloom postfx work. Might have to just do the embers with vfx sprite rather than unity shadergraph.